+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Fixed
 - Fixed duplicated `<code>` in backend form description for noscript-tag.
 
+### Improvements
+- Introduced new `Http\ParameterBag` and `Http\Request`.
+- Removed `filter_input`-usage which causes empty data in various PHP-versions.
+
 
 ## 1.0