+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 ### Improvements
 - Introduced new `Http\ParameterBag` and `Http\Request`.
 - Removed `filter_input`-usage which causes empty data in various PHP-versions.
-
+- Improved description for multisite-field for easier translation.
 
 ## 1.0