+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## 1.1 (NOT RELEASED)
 
+### Updated
+- Updated `chrico/wp-fields` to version `~0.3`.
+
 ### Fixed
 - Fixed duplicated `<code>` in backend form description for noscript-tag.
 
 ### Improvements
+- Updated tests and code according to the new `chrico/wp-fields`-version.
 - Introduced new `Http\ParameterBag` and `Http\Request`.
 - Removed `filter_input`-usage which causes empty data in various PHP-versions.
 - Improved description for multisite-field for easier translation.