+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1 (NOT RELEASED)
 
 ### Updated
+- Updated `readme.txt`.
 - Updated `chrico/wp-fields` to version `~0.3`.
 
 ### Fixed
@@ -13,6 +14,7 @@
 - Introduced new `Http\ParameterBag` and `Http\Request`.
 - Removed `filter_input`-usage which causes empty data in various PHP-versions.
 - Improved description for multisite-field for easier translation.
+- Moved to new [Inpsyde PHP Coding standard](https://github.com/inpsyde/php-coding-standards).
 
 ## 1.0