+++ CHANGELOG.md
@@ -3,16 +3,19 @@
 ## 1.1 (NOT RELEASED)
 
 ### Updated
+- Updated `readme.txt`.
 - Updated `chrico/wp-fields` to version `~0.3`.
 
 ### Fixed
 - Fixed duplicated `<code>` in backend form description for noscript-tag.
 
 ### Improvements
+- Allow saving empty "User > visitorRole" and don't show empty `user.role`.
 - Updated tests and code according to the new `chrico/wp-fields`-version.
 - Introduced new `Http\ParameterBag` and `Http\Request`.
 - Removed `filter_input`-usage which causes empty data in various PHP-versions.
 - Improved description for multisite-field for easier translation.
+- Moved to new [Inpsyde PHP Coding standard](https://github.com/inpsyde/php-coding-standards).
 
 ## 1.0